+++ Forge/src/main/java/com/christofmeg/brutalharvest/common/block/LettuceCropBlock.java
@@ -1,27 +1,14 @@
 package com.christofmeg.brutalharvest.common.block;
 
 import com.christofmeg.brutalharvest.common.init.ItemRegistry;
-import com.christofmeg.brutalharvest.common.item.KnifeItem;
 import net.minecraft.core.BlockPos;
-import net.minecraft.server.level.ServerLevel;
-import net.minecraft.sounds.SoundEvents;
-import net.minecraft.sounds.SoundSource;
-import net.minecraft.util.RandomSource;
-import net.minecraft.world.InteractionHand;
-import net.minecraft.world.InteractionResult;
-import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Items;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.ItemLike;
-import net.minecraft.world.level.Level;
-import net.minecraft.world.level.LevelReader;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.level.block.state.properties.IntegerProperty;
-import net.minecraft.world.level.gameevent.GameEvent;
-import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import org.jetbrains.annotations.NotNull;
@@ -40,40 +27,12 @@ public class LettuceCropBlock extends BaseCropBlock {
 
     public LettuceCropBlock(Properties properties) {
         super(properties);
-        this.registerDefaultState(this.stateDefinition.any().setValue(this.getAgeProperty(), 0));
     }
 
     @Override
-    public @NotNull VoxelShape getShape(@NotNull BlockState state, @NotNull BlockGetter blockGetter, @NotNull BlockPos pos, @NotNull CollisionContext context) {
-        return SHAPE_BY_AGE[this.getAge(state)];
-    }
-
-    @Override
-    public void randomTick(@NotNull BlockState state, @NotNull ServerLevel level, @NotNull BlockPos pos, @NotNull RandomSource random) {
-        int age = state.getValue(AGE);
-        if (age < this.getMaxAge() + 1 && random.nextInt(5) == 0 && level.getRawBrightness(pos.above(), 0) >= 9) {
-            if (age == this.getMaxAge()) {
-                if (random.nextInt(4) == 0) { // 25% chance
-                    state = state.setValue(AGE, age + 1);
-                }
-            } else {
-                state = state.setValue(AGE, age + 1);
-            }
-            state = state.setValue(AGE, age + 1);
-            level.setBlock(pos, state, 2);
-            level.gameEvent(GameEvent.BLOCK_CHANGE, pos, GameEvent.Context.of(state));
-        }
-    }
-
-    @Override
-    public boolean isRandomlyTicking(@NotNull BlockState pState) {
-        return this.getAge(pState) < this.getMaxAge() + 1;
-    }
-
-    @Override
-    public boolean canSurvive(@NotNull BlockState state, @NotNull LevelReader levelReader, BlockPos pos) {
-        BlockPos below = pos.below();
-        return this.mayPlaceOn(levelReader.getBlockState(below), levelReader, below);
+    @NotNull
+    public IntegerProperty getAgeProperty() {
+        return AGE;
     }
 
     @Override
@@ -87,9 +46,8 @@ public class LettuceCropBlock extends BaseCropBlock {
     }
 
     @Override
-    @NotNull
-    public IntegerProperty getAgeProperty() {
-        return AGE;
+    protected ItemStack getBaseItemStack() {
+        return ItemRegistry.LETTUCE.get().getDefaultInstance();
     }
 
     @Override
@@ -98,34 +56,8 @@ public class LettuceCropBlock extends BaseCropBlock {
     }
 
     @Override
-    public boolean isValidBonemealTarget(@NotNull LevelReader levelReader, @NotNull BlockPos pos, @NotNull BlockState state, boolean $$3) {
-        return state.getValue(AGE) < this.getMaxAge();
-    }
-
-    @SuppressWarnings("deprecation")
-    @Override
-    public @NotNull InteractionResult use(@NotNull BlockState state, @NotNull Level level, @NotNull BlockPos pos, @NotNull Player player, @NotNull InteractionHand interactionHand, @NotNull BlockHitResult blockHitResult) {
-        int age = state.getValue(AGE);
-        boolean reachedHarvestAge = age == this.getMaxAge();
-        if (!reachedHarvestAge && player.getItemInHand(interactionHand).is(Items.BONE_MEAL)) {
-            return InteractionResult.PASS;
-        } else if (age > 3) {
-            ItemStack stack = player.getItemInHand(interactionHand);
-            if (stack.getItem() instanceof KnifeItem) {
-                int randomSeeds = level.random.nextInt(2);
-                popResource(level, pos,
-                        age == 4 ? new ItemStack(ItemRegistry.LETTUCE.get()) :
-                                age == 5 ? new ItemStack(this, randomSeeds) :
-                                        ItemStack.EMPTY);
-                level.playSound(null, pos, SoundEvents.ITEM_PICKUP, SoundSource.BLOCKS, 1.0F, 0.8F + level.random.nextFloat() * 0.4F);
-                BlockState newBlockState = state.setValue(AGE, 3);
-                level.setBlock(pos, newBlockState, 2);
-                level.gameEvent(GameEvent.BLOCK_CHANGE, pos, GameEvent.Context.of(player, newBlockState));
-                stack.hurtAndBreak(1, player, (livingEntity) -> livingEntity.broadcastBreakEvent(interactionHand));
-                return InteractionResult.sidedSuccess(level.isClientSide);
-            }
-        }
-        return super.use(state, level, pos, player, interactionHand, blockHitResult);
+    public @NotNull VoxelShape getShape(@NotNull BlockState blockState, @NotNull BlockGetter level, @NotNull BlockPos pos, @NotNull CollisionContext context) {
+        return SHAPE_BY_AGE[this.getAge(blockState)];
     }
 
 }